+++ ephios/plugins/pages/signals.py
@@ -1,0 +1,1 @@
+from django.dispatch import receiver
