+++ ephios/plugins/pages/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+
+from ephios.plugins.basesignup.signup.confirm import RequestConfirmDispositionView
+
+app_name = "plugins"
+urlpatterns = []
