+++ ephios/urls.py
@@ -24,7 +24,7 @@ urlpatterns = [
     path("select2/", include("django_select2.urls")),
     path("", include("ephios.event_management.urls")),
     path("", include("ephios.user_management.urls")),
-    path("", include("ephios.contrib.urls")),
+    path("", include("ephios.plugins.basesignup.urls")),
 ]
 
 if settings.DEBUG:
